+++ example/test/backup_cache_store_test.dart
@@ -1,0 +1,10 @@
+import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
+
+import './common_store_test.dart' as commonStoreTest;
+
+void main() {
+  commonStoreTest.main(
+    'Common Backup store tests',
+    BackupCacheStore(primary: MemCacheStore(), secondary: DbCacheStore()),
+  );
+}